adding to change log
