+++ packages/electoralyze/electoralyze/region/__init__.py
@@ -1,5 +1,5 @@
-from .regions.federal_2021 import Federal2021
+from .regions.federal_2022 import Federal2022
 from .regions.SA1_2021 import SA1_2021
 from .regions.SA2_2021 import SA2_2021
 
-__all__ = ["SA1_2021", "SA2_2021", "Federal2021"]
+__all__ = ["SA1_2021", "SA2_2021", "Federal2022"]
